@@ -10,17 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.11",
-    "@angular/common": "^17.3.11",
-    "@angular/compiler": "^17.3.11",
-    "@angular/core": "^17.3.11",
-    "@angular/forms": "^17.3.11",
-    "@angular/platform-browser": "^17.3.11",
-    "@angular/platform-browser-dynamic": "^17.3.11",
-    "@angular/router": "^17.3.11",
+    "@angular/animations": "^18.0.5",
+    "@angular/common": "^18.0.5",
+    "@angular/compiler": "^18.0.5",
+    "@angular/core": "^18.0.5",
+    "@angular/forms": "^18.0.5",
+    "@angular/platform-browser": "^18.0.5",
+    "@angular/platform-browser-dynamic": "^18.0.5",
+    "@angular/router": "^18.0.5",
     "bootstrap": "^5.3.1",
     "font-awesome": "^4.7.0",
-    "ngx-bootstrap": "^11.0.2",
+    "ngx-bootstrap": "^18.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.7"
@@ -28,7 +28,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.3.8",
     "@angular/cli": "~17.3.8",
-    "@angular/compiler-cli": "^17.3.11",
+    "@angular/compiler-cli": "^18.0.5",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
